<refactor> dense retriever
