+++ retrieval/dense/dense_base.py
@@ -1,10 +1,13 @@
 import os
 import pickle
 import os.path as p
+from itertools import chain
 
 import torch
 import numpy as np
+from torch.utils.data import TensorDataset
 
+from prepare import get_retriever_dataset
 from retrieval.base_retrieval import Retrieval
 
 
@@ -22,7 +25,50 @@ class DenseRetrieval(Retrieval):
         self.encoder_path = p.join(save_dir, f"{self.name}.pth")
 
         self.encoder = None
+        self.tokenizer = None
         self.p_embedding = None
+
+    def _load_dataset(self):
+        # dataset.features : ['query', 'negative_samples', 'label']
+        dataset = get_retriever_dataset(self.args)
+
+        corpus_size = len(dataset["negative_samples"][0])
+        negative_samples = list(chain(*dataset["negative_samples"]))
+
+        # query
+        q_seqs = self.tokenizer(
+            dataset["query"],
+            padding="longest",
+            truncation=True,
+            max_length=self.tokenizer.model_max_length,
+            return_tensors="pt",
+        )
+
+        # negative_samples
+        p_seqs = self.tokenizer(
+            negative_samples,
+            padding="longest",
+            truncation=True,
+            max_length=self.tokenizer.model_max_length,
+            return_tensors="pt",
+        )
+
+        embedding_size = p_seqs.shape[-1]
+
+        for k in p_seqs.keys():
+            p_seqs[k] = p_seqs[k].reshape(-1, corpus_size, embedding_size)
+
+        train_dataset = TensorDataset(
+            p_seqs["input_ids"],
+            p_seqs["attention_mask"],
+            p_seqs["token_type_ids"],
+            q_seqs["input_ids"],
+            q_seqs["attention_mask"],
+            q_seqs["token_type_ids"],
+            torch.tensor(dataset["label"]),
+        )
+
+        return train_dataset
 
     def _load_model(self):
         """학습 때 사용할 모델을 가져옵니다.
@@ -67,7 +113,9 @@ class DenseRetrieval(Retrieval):
         self.encoder.cuda()
 
         with torch.no_grad():
-            q_seqs_val = self.tokenizer(queries, padding="max_length", truncation=True, max_length=512,return_tensors="pt").to("cuda")
+            q_seqs_val = self.tokenizer(
+                queries, padding="max_length", truncation=True, max_length=512, return_tensors="pt"
+            ).to("cuda")
             q_embedding = self.encoder(**q_seqs_val)
             q_embedding.squeeze_()  # in-place
             q_embedding = q_embedding.cpu().detach().numpy()
